Readme & cleanup
