+++ src/nanotron/dataloader.py
@@ -323,68 +323,7 @@ def clm_process(
         desc=f"Grouping texts in chunks of {sequence_length+1}",
     )
     return train_dataset
-
-def vqa_process(
-    raw_dataset: "Dataset",
-    processor: "Idefics3Processor",
-    dataset_processing_num_proc_per_process: int,
-    dataset_overwrite_cache: bool,
-    sequence_length: int,
-):
-    def format_example(example):
-        messages = []
-        for i, x in enumerate(example["en"]):
-            user_message = {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": x["question"]},
-                ]
-            }
-
-            if i == 0:
-                user_message["content"].append(
-                    {"type": "image"},
-                )
-
-            messages.append(user_message)
-            assistant_message = {
-                "role": "assistant",
-                "content": [
-                    {"type": "text", "text": x["answer"]},
-                ]
-            }
-
-            messages.append(assistant_message)
-        return messages
-
-    def _process_examples(examples: Dict, images) -> Dict[str, List[np.ndarray]]:
-        inputs = [
-            processor(
-                text=processor.apply_chat_template(format_example(ex), add_generation_prompt=True),
-                images = [img],
-                return_tensors="np", max_length=sequence_length + 1, padding="longest", truncation=True
-            ) 
-            for ex, img in zip(examples, images)
-        ]
-
-        inputs = {
-            k: [v[k] for v in inputs] for k in ["input_ids", "pixel_values"]
-        }
         
-        return inputs
-
-    train_dataset = raw_dataset.map(
-        _process_examples,
-        input_columns=["qa", "image"],
-        remove_columns=raw_dataset.column_names,
-        batched=True,
-        num_proc=dataset_processing_num_proc_per_process,
-        load_from_cache_file=not dataset_overwrite_cache,
-    )
-
-    return train_dataset
-        
-
 
 # Adapted from: https://github.com/huggingface/transformers/blob/47e1676255e5dd86b9541f734cd4f4bdcbb50f4a/src/transformers/data/data_collator.py#L607
 @dataclasses.dataclass
@@ -460,99 +399,6 @@ class DataCollatorForCLM:
         result = {k: v if isinstance(v, TensorPointer) else torch.from_numpy(v) for k, v in result.items()}
         return result
     
-@dataclasses.dataclass
-class DataCollatorForVQA:
-    sequence_length: int
-    input_pp_rank: int
-    output_pp_rank: int
-    parallel_context: ParallelContext
-    padding_idx: int = 128_002
-
-    def __call__(self, examples: List[Dict[str, List[np.ndarray]]]) -> Dict[str, Union[torch.Tensor, TensorPointer]]:
-        # Process the case when current rank doesn't require data. We return `TensorPointer` that points to ranks having the data.
-        current_pp_rank = dist.get_rank(self.parallel_context.pp_pg)
-        if current_pp_rank not in [
-            self.input_pp_rank,
-            self.output_pp_rank,
-        ]:
-            assert all(len(example) == 0 for example in examples)
-            return {
-                "input_ids": TensorPointer(group_rank=self.input_pp_rank),
-                "input_mask": TensorPointer(group_rank=self.input_pp_rank),
-                "label_ids": TensorPointer(group_rank=self.output_pp_rank),
-                "label_mask": TensorPointer(group_rank=self.output_pp_rank),
-                "pixel_values": TensorPointer(group_rank=self.input_pp_rank),
-            }
-
-        # Make sure we load only what's necessary, ie we only load `input_ids` and `pixel_values` column.
-        assert all(list(example.keys()) == ["input_ids", "pixel_values"] for example in examples)
-
-        max_n_patches = max([len(examples[i]["pixel_values"][0]) for i in range(len(examples))])
-
-        padded_pixel_values = []
-
-        for example in examples:
-            pixel_values = example["pixel_values"]
-            current_patches = len(pixel_values[0])
-            
-            # Pad the pixel_values to have max_n_patches along dimension 1 (patches)
-            padding = ((0, 0), (0, max_n_patches - current_patches), (0, 0), (0, 0), (0, 0))  # Only pad the patches dimension
-            padded_values = np.pad(pixel_values, pad_width=padding, mode='constant', constant_values=0)
-            padded_pixel_values.append(padded_values)
-            
-
-        # Step 3: Stack padded pixel_values and pixel_attention_masks
-        pixel_values = np.vstack(padded_pixel_values)  # Stacked pixel_values
-        result: Dict[str, Union[np.ndarray, TensorPointer]] = {}
-
-        result["input_ids"] = TensorPointer(group_rank=self.input_pp_rank)
-        result["input_mask"] = TensorPointer(group_rank=self.input_pp_rank)
-        result["label_ids"] = TensorPointer(group_rank=self.output_pp_rank)
-        result["label_mask"] = TensorPointer(group_rank=self.output_pp_rank)
-        result["pixel_values"] = TensorPointer(group_rank=self.input_pp_rank)
-
-        def pad_tokens(inputs = True):
-            padded_tokens = []
-            token_masks = []
-
-            max_seq_length = max([len(examples[i]["input_ids"][0]) for i in range(len(examples))]) - 1
-            # make it divisible by 4 for tp
-            max_seq_length = max_seq_length + (4 - max_seq_length % 4) % 4
-
-            for example in examples:
-                input_ids = example["input_ids"]
-                if type(input_ids) == list:
-                    input_ids = np.array(input_ids)
-
-                if inputs:
-                    input_ids = input_ids[:, :-1]
-                else:
-                    input_ids = input_ids[:, 1:]
-
-                current_length = input_ids.shape[1]
-
-                padding = ((0, 0), (0, max_seq_length - current_length))
-                input_ids = np.pad(input_ids, pad_width=padding, mode='constant', constant_values=self.padding_idx)
-                padded_tokens.append(input_ids)
-
-                mask = np.ones((1, current_length), dtype=np.bool_)
-                mask = np.pad(mask, pad_width=padding, mode='constant', constant_values=0)
-                token_masks.append(mask)
-
-            padded_tokens = np.vstack(padded_tokens)
-            token_masks = np.vstack(token_masks)
-
-            return padded_tokens, token_masks
-
-        if current_pp_rank == self.input_pp_rank:
-            result["input_ids"], result["input_mask"] = pad_tokens(inputs=True)
-            result["pixel_values"] = pixel_values
-        
-        if current_pp_rank == self.output_pp_rank:
-            result["label_ids"], result["label_mask"] = pad_tokens(inputs=False)
-
-        result = {k: v if isinstance(v, TensorPointer) else torch.from_numpy(v) for k, v in result.items()}
-        return result
 
 # Adapted from https://github.com/huggingface/transformers/blob/47e1676255e5dd86b9541f734cd4f4bdcbb50f4a/src/transformers/trainer.py#L763-L835
 def get_sampler(
@@ -607,7 +453,8 @@ def get_train_dataloader(
     dataloader_drop_last: bool = True,
     dataloader_pin_memory: bool = True,
     use_loop_to_round_batch_size: bool = False,
-    dataset_columns = ["input_ids"]
+    dataset_columns = ["input_ids"],
+    collator_builder=None
 ) -> DataLoader:
     if not isinstance(train_dataset, datasets.Dataset):
         raise ValueError(f"training requires a datasets.Dataset, but got {type(train_dataset)}")
@@ -636,8 +483,8 @@ def get_train_dataloader(
         # No need to spawn a lot of workers, we can just use main
         dataloader_num_workers = 0
 
-    if "pixel_values" in dataset_columns:
-        data_collator = DataCollatorForVQA(
+    if collator_builder is not None:
+        data_collator = collator_builder(
             sequence_length=sequence_length,
             input_pp_rank=input_pp_rank,
             output_pp_rank=output_pp_rank,
@@ -650,7 +497,7 @@ def get_train_dataloader(
             output_pp_rank=output_pp_rank,
             parallel_context=parallel_context,
         )
-
+        
     # Compute size and rank of dataloader workers
     dp_ranks_size = parallel_context.dp_pg.size()
     dp_rank = parallel_context.dp_pg.rank()
