+++ src/nanotron/modular_dataloader/iterable.py
@@ -1,14 +1,16 @@
-from typing import Any, Dict, List, Optional
-from datasets import IterableDataset
-from nanotron.parallel.context import ParallelContext
-from nanotron import distributed as dist
-from nanotron.modular_dataloader.base import SampleEncoder, BatchEncoder, T_encoded_sample
+import itertools
 from multiprocessing.pool import ThreadPool
+from typing import Any, Dict, List
+
+from datasets import IterableDataset
 from datasets.distributed import split_dataset_by_node
 from torch.utils.data import DataLoader
 
+from nanotron import distributed as dist
+from nanotron.modular_dataloader.base import BatchEncoder, SampleEncoder
+from nanotron.parallel.context import ParallelContext
+from nanotron.parallel.pipeline_parallel.tensor_pointer import TensorPointer
 
-from nanotron.parallel.pipeline_parallel.tensor_pointer import TensorPointer  
 
 def from_columns(batch: Dict[str, List]):
     return [{k: batch[k][i] for k in batch} for i in range(len(batch[list(batch.keys())[0]]))]
@@ -20,7 +22,7 @@ class EmptyIterableDataset(IterableDataset):
     def __len__(self):
         return 0
 
-import itertools
+
 
 class EmptyDataset(IterableDataset):
     def __init__(self, input_pp_rank: int, output_pp_rank: int, num_shards: int):
@@ -59,10 +61,10 @@ def get_train_dataloader(
     sample_encoding_workers: int,
     batch_encoding_workers: int,
     drop_last: bool = True,
-): 
+):
     if not isinstance(train_dataset, IterableDataset):
         raise ValueError("Dataset should be a datasets.IterableDataset")
-    
+
     if dist.get_rank(parallel_context.pp_pg) not in [input_pp_rank, output_pp_rank]:
 
         def generator():
@@ -83,7 +85,7 @@ def get_train_dataloader(
             num_workers=0,
             collate_fn=lambda x: x[0],
         )
-    
+
     train_dataset = split_dataset_by_node(train_dataset, rank=parallel_context.dp_pg.rank(), world_size=parallel_context.dp_pg.size())
     train_dataset = train_dataset.shuffle(seed=seed_worker)
 
@@ -94,14 +96,14 @@ def get_train_dataloader(
             encoded_batch = sample_worker_pool.map(sample_encoder.encode, batch)
 
         return {"sample_encoded": encoded_batch}
-    
+
     train_dataset = train_dataset.map(encode_samples_batched, batched=True, remove_columns=train_dataset.column_names, batch_size=sample_encoding_batch)
-    
-    
+
+
     def collate_fn(batch: List[Dict[str, Any]]):
         batch = [x["sample_encoded"] for x in batch]
         return batch_encoder.encode(batch)
-    
+
     if consumed_train_samples > 0:
         dp_size = parallel_context.dp_pg.size()
         skip_batches = consumed_train_samples // dp_size
@@ -120,5 +122,5 @@ def get_train_dataloader(
 
 
 
-    
-    
+
+
